拡張をファイルに切り出す
